# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Shelter API
  description: OpenAPI specification for the Shelter rental marketplace backend.
  version: 1.0.0
  contact:
    name: Ibrahim Abdullaziz
    email: ibrahimabdullaziz55@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:2005
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request payload or parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Invalid credentials
    User:
      type: object
      required:
        - id
        - email
        - firstName
        - lastName
        - role
      properties:
        id:
          type: string
          format: uuid
          example: 7d08f7c4-9d6b-4e12-bfdf-0f8e0b6f0ee8
        email:
          type: string
          format: email
          example: guest@example.com
        firstName:
          type: string
          example: Jane
        lastName:
          type: string
          example: Doe
        role:
          type: string
          enum: [GUEST, HOST, ADMIN]
          example: GUEST
        isVerified:
          type: boolean
          example: true
...
    VerifyEmailRequest:
      type: object
      required:
        - email
        - code
      properties:
        email:
          type: string
          format: email
          example: guest@example.com
        code:
          type: string
          minLength: 6
          maxLength: 6
          example: "123456"
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: guest@example.com
    ResetPasswordRequest:
      type: object
      required:
        - email
        - code
        - password
      properties:
        email:
          type: string
          format: email
          example: guest@example.com
        code:
          type: string
          minLength: 6
          maxLength: 6
          example: "123456"
        password:
          type: string
          format: password
          minLength: 8
          example: NewStrongPass123
...
paths:
...
  /api/auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify user email
      description: Confirm a user email address using the 6-digit OTP code generated during signup.
      operationId: verifyEmail
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailRequest"
            example:
              email: guest@example.com
              code: "123456"
      responses:
        "200":
          description: Email verified successfully
          content:
            application/json:
              schema:
                type: object
                required: [status, message, data]
                properties:
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: Email verified successfully
                  data:
                    $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset code
      description: Send a password reset OTP to the user's email.
      operationId: forgotPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
            example:
              email: guest@example.com
      responses:
        "200":
          description: Password reset code sent successfully
          content:
            application/json:
              schema:
                type: object
                required: [status, message, data]
                properties:
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: Password changed successfully
                  data:
                    type: object
                    nullable: true
                    example: null
        "404":
          $ref: "#/components/responses/NotFound"
  /api/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password
      description: Reset a user's password using the email and verification code.
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
            example:
              email: guest@example.com
              code: "123456"
              password: NewStrongPass123
      responses:
        "200":
          description: Password updated successfully
          content:
            application/json:
              schema:
                type: object
                required: [status, message, data]
                properties:
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: Password changed successfully
                  data:
                    $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...